+++ app/src/main/java/ru/aevd/androidacademymovieapp/OnMoviesItemClicked.kt
@@ -1,5 +1,5 @@
 package ru.aevd.androidacademymovieapp
 
 interface OnMoviesItemClicked {
-    fun onClick(movie_id: Long)
+    fun onClick(movie_id: Int)
 }
