+++ app/src/main/java/ru/aevd/androidacademymovieapp/TransactionsFragmentClicks.kt
@@ -1,6 +1,6 @@
 package ru.aevd.androidacademymovieapp
 
 interface TransactionsFragmentClicks {
-    fun showMovieDetails(movie_id: Long)
+    fun showMovieDetails(movie_id: Int)
     fun navigateBack()
 }
